Tree Generator added but not completely working
Added the tree generator class.
Modified the code to work with glm functions.
Added a tree shader.  There is a warning for using the texture command.
This is what I used before and didn't have a problem.  I did attempt to
use GLSL 1.30 instead with the same results i.e. no tree being rendered.
Tree generator has a render function.
Looks like everything is hooked up to render and I checked that the
shader is getting the information via glslDevil.